build and setup board for v4.1/4.1
