+++ firmware/real_machines/ARWAC-v4.1.md
@@ -1,0 +1,4 @@
+ESPBTMACADDRESS 10:52:1c:76:30:42 - 4.1
+ESPBTMACADDRESS 24:d7:eb:0e:dc:da - 4.2
+ESPBTMACADDRESS FC:F5:C4:00:FA:C6 - SCOOTERBOT
+
