+++ firmware/stupid-mobility-adapter-v3/.vscode/arduino.json
@@ -1,0 +1,6 @@
+{
+    "configuration": "FlashFreq=80,UploadSpeed=921600,DebugLevel=none",
+    "board": "esp32:esp32:esp32doit-devkit-v1",
+    "port": "COM3",
+    "sketch": "stupid-mobility-adapter-v3.ino"
+}
