--- firmware/stupid-mobility-adapter-v3/rcinput.h
@@ -1,14 +1,0 @@
-#ifndef RCINPUT_H
-#define RCINPUT_H
-
-//extern volatile int pwm_value_throttle;
-//extern volatile int pwm_value_steer;
-
-//void fallingThrottle();
-//void risingThrottle();
-//void fallingSteer();
-//void risingSteer();
-void startRCInput();
-void doRCInput();
-
-#endif
