+++ stupid-mobility-adapter-v3/stupid-mobility-adapter-v3/Manufacturing/stupid-mobility-adapter-v3-B_Silkscreen.gbr
@@ -1,0 +1,15 @@
+%TF.GenerationSoftware,KiCad,Pcbnew,(6.0.2)*%
+%TF.CreationDate,2022-11-16T14:07:38+00:00*%
+%TF.ProjectId,stupid-mobility-adapter-v3,73747570-6964-42d6-9d6f-62696c697479,rev?*%
+%TF.SameCoordinates,Original*%
+%TF.FileFunction,Legend,Bot*%
+%TF.FilePolarity,Positive*%
+%FSLAX46Y46*%
+G04 Gerber Fmt 4.6, Leading zero omitted, Abs format (unit mm)*
+G04 Created by KiCad (PCBNEW (6.0.2)) date 2022-11-16 14:07:38*
+%MOMM*%
+%LPD*%
+G01*
+G04 APERTURE LIST*
+G04 APERTURE END LIST*
+M02*
