+++ stupid-mobility-adapter-v3/stupid-mobility-adapter-v3/Manufacturing/stupid-mobility-adapter-v3-Edge_Cuts.gbr
@@ -1,0 +1,26 @@
+%TF.GenerationSoftware,KiCad,Pcbnew,(6.0.2)*%
+%TF.CreationDate,2022-11-16T14:07:38+00:00*%
+%TF.ProjectId,stupid-mobility-adapter-v3,73747570-6964-42d6-9d6f-62696c697479,rev?*%
+%TF.SameCoordinates,Original*%
+%TF.FileFunction,Profile,NP*%
+%FSLAX46Y46*%
+G04 Gerber Fmt 4.6, Leading zero omitted, Abs format (unit mm)*
+G04 Created by KiCad (PCBNEW (6.0.2)) date 2022-11-16 14:07:38*
+%MOMM*%
+%LPD*%
+G01*
+G04 APERTURE LIST*
+%TA.AperFunction,Profile*%
+%ADD10C,0.100000*%
+%TD*%
+G04 APERTURE END LIST*
+D10*
+X255778000Y-52070000D02*
+X160837500Y-52070000D01*
+X160837500Y-52070000D02*
+X160837500Y-145532500D01*
+X160837500Y-145532500D02*
+X255778000Y-145532500D01*
+X255778000Y-145532500D02*
+X255778000Y-52070000D01*
+M02*
