+++ stupid-mobility-adapter-v3/stupid-mobility-adapter-v3/Manufacturing/stupid-mobility-adapter-v3-F_Mask.gbr
@@ -1,0 +1,374 @@
+%TF.GenerationSoftware,KiCad,Pcbnew,(6.0.2)*%
+%TF.CreationDate,2022-11-16T14:07:38+00:00*%
+%TF.ProjectId,stupid-mobility-adapter-v3,73747570-6964-42d6-9d6f-62696c697479,rev?*%
+%TF.SameCoordinates,Original*%
+%TF.FileFunction,Soldermask,Top*%
+%TF.FilePolarity,Negative*%
+%FSLAX46Y46*%
+G04 Gerber Fmt 4.6, Leading zero omitted, Abs format (unit mm)*
+G04 Created by KiCad (PCBNEW (6.0.2)) date 2022-11-16 14:07:38*
+%MOMM*%
+%LPD*%
+G01*
+G04 APERTURE LIST*
+G04 Aperture macros list*
+%AMRoundRect*
+0 Rectangle with rounded corners*
+0 $1 Rounding radius*
+0 $2 $3 $4 $5 $6 $7 $8 $9 X,Y pos of 4 corners*
+0 Add a 4 corners polygon primitive as box body*
+4,1,4,$2,$3,$4,$5,$6,$7,$8,$9,$2,$3,0*
+0 Add four circle primitives for the rounded corners*
+1,1,$1+$1,$2,$3*
+1,1,$1+$1,$4,$5*
+1,1,$1+$1,$6,$7*
+1,1,$1+$1,$8,$9*
+0 Add four rect primitives between the rounded corners*
+20,1,$1+$1,$2,$3,$4,$5,0*
+20,1,$1+$1,$4,$5,$6,$7,0*
+20,1,$1+$1,$6,$7,$8,$9,0*
+20,1,$1+$1,$8,$9,$2,$3,0*%
+G04 Aperture macros list end*
+%ADD10C,1.600000*%
+%ADD11O,1.600000X1.600000*%
+%ADD12R,2.600000X2.600000*%
+%ADD13C,2.600000*%
+%ADD14R,1.800000X1.800000*%
+%ADD15C,1.800000*%
+%ADD16R,1.700000X1.700000*%
+%ADD17O,1.700000X1.700000*%
+%ADD18R,1.050000X1.500000*%
+%ADD19O,1.050000X1.500000*%
+%ADD20R,2.400000X2.400000*%
+%ADD21C,2.400000*%
+%ADD22C,5.600000*%
+%ADD23R,1.600000X1.600000*%
+%ADD24O,1.200000X2.000000*%
+%ADD25RoundRect,0.300000X-0.300000X-0.700000X0.300000X-0.700000X0.300000X0.700000X-0.300000X0.700000X0*%
+%ADD26RoundRect,0.300000X0.300000X0.700000X-0.300000X0.700000X-0.300000X-0.700000X0.300000X-0.700000X0*%
+%ADD27C,1.524000*%
+%ADD28RoundRect,0.250000X-0.600000X0.600000X-0.600000X-0.600000X0.600000X-0.600000X0.600000X0.600000X0*%
+%ADD29C,1.700000*%
+%ADD30O,8.000000X3.800000*%
+%ADD31R,1.524000X1.524000*%
+G04 APERTURE END LIST*
+D10*
+%TO.C,R4*%
+X201676000Y-69596000D03*
+D11*
+X201676000Y-77216000D03*
+%TD*%
+D10*
+%TO.C,R8*%
+X226568000Y-70612000D03*
+D11*
+X239268000Y-70612000D03*
+%TD*%
+D12*
+%TO.C,U3*%
+X249474000Y-98049000D03*
+D13*
+X249474000Y-92969000D03*
+X249474000Y-87889000D03*
+X249474000Y-82809000D03*
+X249474000Y-77729000D03*
+X249474000Y-72649000D03*
+X249474000Y-67569000D03*
+%TD*%
+D10*
+%TO.C,R10*%
+X236982000Y-113030000D03*
+D11*
+X236982000Y-120650000D03*
+%TD*%
+D14*
+%TO.C,D4*%
+X240538000Y-108971000D03*
+D15*
+X240538000Y-106431000D03*
+%TD*%
+D10*
+%TO.C,R14*%
+X226187000Y-106553000D03*
+D11*
+X226187000Y-114173000D03*
+%TD*%
+D16*
+%TO.C,J10*%
+X213868000Y-92710000D03*
+D17*
+X213868000Y-90170000D03*
+%TD*%
+D10*
+%TO.C,R13*%
+X248920000Y-113030000D03*
+D11*
+X248920000Y-120650000D03*
+%TD*%
+D18*
+%TO.C,Q2*%
+X251714000Y-125222000D03*
+D19*
+X250444000Y-125222000D03*
+X249174000Y-125222000D03*
+%TD*%
+D16*
+%TO.C,J9*%
+X165100000Y-85598000D03*
+D17*
+X165100000Y-83058000D03*
+X165100000Y-80518000D03*
+%TD*%
+D10*
+%TO.C,R5*%
+X165100000Y-89408000D03*
+D11*
+X172720000Y-89408000D03*
+%TD*%
+D10*
+%TO.C,R9*%
+X231648000Y-63481000D03*
+D11*
+X218948000Y-63481000D03*
+%TD*%
+D14*
+%TO.C,D5*%
+X248920000Y-108971000D03*
+D15*
+X248920000Y-106431000D03*
+%TD*%
+D20*
+%TO.C,J2*%
+X192111250Y-140970000D03*
+D21*
+X188611250Y-140970000D03*
+X185111250Y-140970000D03*
+%TD*%
+D22*
+%TO.C,3.2*%
+X251714000Y-141732000D03*
+%TD*%
+D16*
+%TO.C,J6*%
+X188722000Y-96012000D03*
+D17*
+X188722000Y-93472000D03*
+%TD*%
+D14*
+%TO.C,D3*%
+X238506000Y-63500000D03*
+D15*
+X235966000Y-63500000D03*
+%TD*%
+D16*
+%TO.C,J7*%
+X235458000Y-85344000D03*
+D17*
+X235458000Y-87884000D03*
+X235458000Y-90424000D03*
+X235458000Y-92964000D03*
+%TD*%
+D16*
+%TO.C,J8*%
+X188722000Y-88397000D03*
+D17*
+X188722000Y-85857000D03*
+%TD*%
+D23*
+%TO.C,U4*%
+X202194000Y-65522000D03*
+D11*
+X204734000Y-65522000D03*
+X207274000Y-65522000D03*
+X209814000Y-65522000D03*
+X209814000Y-57902000D03*
+X207274000Y-57902000D03*
+X204734000Y-57902000D03*
+X202194000Y-57902000D03*
+%TD*%
+D22*
+%TO.C,3.2*%
+X164465000Y-55753000D03*
+%TD*%
+D10*
+%TO.C,R1*%
+X169164000Y-64008000D03*
+D11*
+X169164000Y-71628000D03*
+%TD*%
+D14*
+%TO.C,D1*%
+X235961000Y-78213000D03*
+D15*
+X238501000Y-78213000D03*
+%TD*%
+D10*
+%TO.C,R3*%
+X172720000Y-92710000D03*
+D11*
+X165100000Y-92710000D03*
+%TD*%
+D10*
+%TO.C,R11*%
+X252476000Y-113030000D03*
+D11*
+X252476000Y-120650000D03*
+%TD*%
+D10*
+%TO.C,R7*%
+X219710000Y-78213000D03*
+D11*
+X232410000Y-78213000D03*
+%TD*%
+D24*
+%TO.C,U5*%
+X216002670Y-127873680D03*
+X213462670Y-127873680D03*
+X210922670Y-127873680D03*
+X208382670Y-127873680D03*
+X205842670Y-127873680D03*
+X203302670Y-127873680D03*
+X200762670Y-127873680D03*
+X198222670Y-127873680D03*
+X195682670Y-127873680D03*
+X193142670Y-127873680D03*
+X190602670Y-127873680D03*
+X188062670Y-127873680D03*
+X185522670Y-127873680D03*
+X182982670Y-127873680D03*
+X180442670Y-127873680D03*
+X177902670Y-127873680D03*
+X175362670Y-127873680D03*
+X172822670Y-127873680D03*
+D25*
+X170285390Y-127870000D03*
+X170277230Y-102470000D03*
+D24*
+X172946950Y-102470000D03*
+X175486950Y-102470000D03*
+X178026950Y-102470000D03*
+X180566950Y-102470000D03*
+X183106950Y-102470000D03*
+X185646950Y-102470000D03*
+X188186950Y-102470000D03*
+X190726950Y-102470000D03*
+X193266950Y-102470000D03*
+X195806950Y-102470000D03*
+X198346950Y-102470000D03*
+X200886950Y-102470000D03*
+X203426950Y-102470000D03*
+X205966950Y-102470000D03*
+X208379950Y-102470000D03*
+X210919950Y-102470000D03*
+X213459950Y-102470000D03*
+D26*
+X215999950Y-102470000D03*
+%TD*%
+D20*
+%TO.C,J4*%
+X204795000Y-140970000D03*
+D21*
+X201295000Y-140970000D03*
+X197795000Y-140970000D03*
+%TD*%
+D14*
+%TO.C,D2*%
+X219959000Y-70612000D03*
+D15*
+X222499000Y-70612000D03*
+%TD*%
+D27*
+%TO.C,U1*%
+X194828000Y-85354000D03*
+X194828000Y-87894000D03*
+X194828000Y-90434000D03*
+X194828000Y-92974000D03*
+X194828000Y-95514000D03*
+X194828000Y-98054000D03*
+%TD*%
+D22*
+%TO.C,3.2*%
+X164592000Y-141605000D03*
+%TD*%
+D10*
+%TO.C,R2*%
+X165100000Y-71755000D03*
+D11*
+X165100000Y-64135000D03*
+%TD*%
+D20*
+%TO.C,J5*%
+X178942500Y-141087500D03*
+D21*
+X175442500Y-141087500D03*
+%TD*%
+D28*
+%TO.C,J1*%
+X221615000Y-138312500D03*
+D29*
+X221615000Y-140852500D03*
+X219075000Y-138312500D03*
+X219075000Y-140852500D03*
+X216535000Y-138312500D03*
+X216535000Y-140852500D03*
+X213995000Y-138312500D03*
+X213995000Y-140852500D03*
+%TD*%
+D28*
+%TO.C,J3*%
+X241300000Y-138312500D03*
+D29*
+X241300000Y-140852500D03*
+X238760000Y-138312500D03*
+X238760000Y-140852500D03*
+X236220000Y-138312500D03*
+X236220000Y-140852500D03*
+X233680000Y-138312500D03*
+X233680000Y-140852500D03*
+%TD*%
+D10*
+%TO.C,R6*%
+X207010000Y-69596000D03*
+D11*
+X207010000Y-77216000D03*
+%TD*%
+D18*
+%TO.C,Q1*%
+X237744000Y-125476000D03*
+D19*
+X239014000Y-125476000D03*
+X240284000Y-125476000D03*
+%TD*%
+D22*
+%TO.C,3.2*%
+X251968000Y-55753000D03*
+%TD*%
+D30*
+%TO.C,F1*%
+X225728000Y-56134000D03*
+X235028000Y-56134000D03*
+%TD*%
+D31*
+%TO.C,U2*%
+X176458500Y-57639000D03*
+X173918500Y-57639000D03*
+X184586500Y-57639000D03*
+X187126500Y-57639000D03*
+X184586500Y-76181000D03*
+X187126500Y-76181000D03*
+X176458500Y-76181000D03*
+X173918500Y-76181000D03*
+%TD*%
+D10*
+%TO.C,R12*%
+X240538000Y-113030000D03*
+D11*
+X240538000Y-120650000D03*
+%TD*%
+D23*
+%TO.C,BZ1*%
+X227330000Y-101092000D03*
+D10*
+X227330000Y-98092000D03*
+%TD*%
+M02*
