+++ stupid-mobility-adapter-v3/stupid-mobility-adapter-v3/Manufacturing/stupid-mobility-adapter-v3-NPTH.drl
@@ -1,0 +1,13 @@
+M48
+; DRILL file {KiCad (6.0.2)} date Wed Nov 16 14:08:11 2022
+; FORMAT={-:-/ absolute / inch / decimal}
+; #@! TF.CreationDate,2022-11-16T14:08:11+00:00
+; #@! TF.GenerationSoftware,Kicad,Pcbnew,(6.0.2)
+; #@! TF.FileFunction,NonPlated,1,2,NPTH
+FMAT,2
+INCH
+%
+G90
+G05
+T0
+M30
