+++ stupid-mobility-adapter-v3/stupid-mobility-adapter-v3/Manufacturing/stupid-mobility-adapter-v3-PTH.drl
@@ -1,0 +1,204 @@
+M48
+; DRILL file {KiCad (6.0.2)} date Wed Nov 16 14:08:11 2022
+; FORMAT={-:-/ absolute / inch / decimal}
+; #@! TF.CreationDate,2022-11-16T14:08:11+00:00
+; #@! TF.GenerationSoftware,Kicad,Pcbnew,(6.0.2)
+; #@! TF.FileFunction,Plated,1,2,PTH
+FMAT,2
+INCH
+; #@! TA.AperFunction,Plated,PTH,ComponentDrill
+T1C0.0295
+; #@! TA.AperFunction,Plated,PTH,ComponentDrill
+T2C0.0300
+; #@! TA.AperFunction,Plated,PTH,ComponentDrill
+T3C0.0315
+; #@! TA.AperFunction,Plated,PTH,ComponentDrill
+T4C0.0354
+; #@! TA.AperFunction,Plated,PTH,ComponentDrill
+T5C0.0394
+; #@! TA.AperFunction,Plated,PTH,ComponentDrill
+T6C0.0472
+; #@! TA.AperFunction,Plated,PTH,ComponentDrill
+T7C0.0512
+; #@! TA.AperFunction,Plated,PTH,ComponentDrill
+T8C0.0787
+; #@! TA.AperFunction,Plated,PTH,ComponentDrill
+T9C0.1260
+%
+G90
+G05
+T1
+X9.36Y-4.94
+X9.41Y-4.94
+X9.46Y-4.94
+X9.81Y-4.93
+X9.86Y-4.93
+X9.91Y-4.93
+T2
+X6.8472Y-2.2693
+X6.8472Y-2.9993
+X6.9472Y-2.2693
+X6.9472Y-2.9993
+X7.2672Y-2.2693
+X7.2672Y-2.9993
+X7.3672Y-2.2693
+X7.3672Y-2.9993
+X7.6704Y-3.3604
+X7.6704Y-3.4604
+X7.6704Y-3.5604
+X7.6704Y-3.6604
+X7.6704Y-3.7604
+X7.6704Y-3.8604
+T3
+X6.5Y-2.525
+X6.5Y-2.825
+X6.5Y-3.52
+X6.5Y-3.65
+X6.66Y-2.52
+X6.66Y-2.82
+X6.7038Y-4.0343
+X6.7041Y-5.0343
+X6.8Y-3.52
+X6.8Y-3.65
+X6.804Y-5.0344
+X6.8089Y-4.0343
+X6.904Y-5.0344
+X6.9089Y-4.0343
+X7.004Y-5.0344
+X7.0089Y-4.0343
+X7.104Y-5.0344
+X7.1089Y-4.0343
+X7.204Y-5.0344
+X7.2089Y-4.0343
+X7.304Y-5.0344
+X7.3089Y-4.0343
+X7.404Y-5.0344
+X7.4089Y-4.0343
+X7.504Y-5.0344
+X7.5089Y-4.0343
+X7.604Y-5.0344
+X7.6089Y-4.0343
+X7.704Y-5.0344
+X7.7089Y-4.0343
+X7.804Y-5.0344
+X7.8089Y-4.0343
+X7.904Y-5.0344
+X7.9089Y-4.0343
+X7.94Y-2.74
+X7.94Y-3.04
+X7.9604Y-2.2796
+X7.9604Y-2.5796
+X8.004Y-5.0344
+X8.0089Y-4.0343
+X8.0604Y-2.2796
+X8.0604Y-2.5796
+X8.104Y-5.0344
+X8.1089Y-4.0343
+X8.15Y-2.74
+X8.15Y-3.04
+X8.1604Y-2.2796
+X8.1604Y-2.5796
+X8.2039Y-4.0343
+X8.204Y-5.0344
+X8.2604Y-2.2796
+X8.2604Y-2.5796
+X8.3039Y-4.0343
+X8.304Y-5.0344
+X8.4039Y-4.0343
+X8.404Y-5.0344
+X8.5039Y-4.0343
+X8.504Y-5.0344
+X8.62Y-2.4993
+X8.65Y-3.0793
+X8.905Y-4.195
+X8.905Y-4.495
+X8.92Y-2.78
+X9.12Y-2.4993
+X9.15Y-3.0793
+X9.33Y-4.45
+X9.33Y-4.75
+X9.42Y-2.78
+X9.47Y-4.45
+X9.47Y-4.75
+X9.8Y-4.45
+X9.8Y-4.75
+X9.94Y-4.45
+X9.94Y-4.75
+T4
+X8.6598Y-2.78
+X8.7598Y-2.78
+X9.2898Y-3.0793
+X9.29Y-2.5
+X9.3898Y-3.0793
+X9.39Y-2.5
+X9.47Y-4.1902
+X9.47Y-4.2902
+X9.8Y-4.1902
+X9.8Y-4.2902
+T5
+X6.5Y-3.17
+X6.5Y-3.27
+X6.5Y-3.37
+X7.43Y-3.3802
+X7.43Y-3.4802
+X7.43Y-3.68
+X7.43Y-3.78
+X8.42Y-3.55
+X8.42Y-3.65
+X8.425Y-5.4454
+X8.425Y-5.5454
+X8.525Y-5.4454
+X8.525Y-5.5454
+X8.625Y-5.4454
+X8.625Y-5.5454
+X8.725Y-5.4454
+X8.725Y-5.5454
+X8.95Y-3.8619
+X8.95Y-3.98
+X9.2Y-5.4454
+X9.2Y-5.5454
+X9.27Y-3.36
+X9.27Y-3.46
+X9.27Y-3.56
+X9.27Y-3.66
+X9.3Y-5.4454
+X9.3Y-5.5454
+X9.4Y-5.4454
+X9.4Y-5.5454
+X9.5Y-5.4454
+X9.5Y-5.5454
+T6
+X6.9072Y-5.5546
+X7.045Y-5.5546
+X7.2878Y-5.55
+X7.4256Y-5.55
+X7.5634Y-5.55
+X7.7872Y-5.55
+X7.925Y-5.55
+X8.0628Y-5.55
+T7
+X9.8218Y-2.6602
+X9.8218Y-2.8602
+X9.8218Y-3.0602
+X9.8218Y-3.2602
+X9.8218Y-3.4602
+X9.8218Y-3.6602
+X9.8218Y-3.8602
+T9
+X6.475Y-2.195
+X6.48Y-5.575
+X9.91Y-5.58
+X9.92Y-2.195
+T8
+G00X8.8082Y-2.21
+M15
+G01X8.9657Y-2.21
+M16
+G05
+G00X9.1743Y-2.21
+M15
+G01X9.3318Y-2.21
+M16
+G05
+T0
+M30
