+++ stupid-mobility-adapter-v3/stupid-mobility-adapter-v3/fp-info-cache
@@ -1,4 +1,4 @@
-19894509736235760
+19896178892947507
 Audio_Module
 Reverb_BTDR-1H
 Digital Reverberation Unit, http://www.belton.co.kr/inc/downfile.php?seq=17&file=pdf (footprint from http://www.uk-electronic.de/PDF/BTDR-1.pdf)
@@ -84890,6 +84890,13 @@ MP1584
 4
 stuipid-electronics-components
 MPC4725-Sparkfun
+
+
+0
+6
+6
+stuipid-electronics-components
+TJA1050-CAN-TRANS-MODULE
 
 
 0
