+++ stupid-mobility-adapter-v3/stupid-mobility-adapter-v3/stupid-mobility-adapter-v3.kicad_prl
@@ -1,6 +1,6 @@
 {
   "board": {
-    "active_layer": 37,
+    "active_layer": 0,
     "active_layer_preset": "",
     "auto_track_width": false,
     "hidden_nets": [],
@@ -61,7 +61,7 @@
       38
     ],
     "visible_layers": "ffdffff_ffffffff",
-    "zone_display_mode": 1
+    "zone_display_mode": 0
   },
   "meta": {
     "filename": "stupid-mobility-adapter-v3.kicad_prl",
