+++ stupid-mobility-adapter-v3/stupid-mobility-adapter-v3/stupid-mobility-adapter-v3.kicad_sch
@@ -1681,7 +1681,6 @@
   (no_connect (at 245.11 132.08) (uuid b00b01b2-286f-4ffb-9299-6d2a904b542d))
   (no_connect (at 50.8 34.29) (uuid bbeb4758-2d88-449c-a139-6d130c88e4f6))
   (no_connect (at 201.93 147.32) (uuid c73ca981-71cf-4f32-b21e-3aa18b0f2771))
-  (no_connect (at 199.39 121.92) (uuid cb6d2e9d-65b9-4099-afdf-888279b3a911))
   (no_connect (at 102.87 35.56) (uuid f2065ece-4383-403a-b7e6-75a596d44ea4))
   (no_connect (at 90.17 35.56) (uuid f2065ece-4383-403a-b7e6-75a596d44ea5))
   (no_connect (at 242.57 147.32) (uuid f6522f87-6c3b-41c4-a85f-1074142c4d3f))
@@ -2294,6 +2293,10 @@
     (effects (font (size 1.27 1.27)) (justify left bottom))
     (uuid 3ab9fe19-d84d-4278-bf15-74ba2c6599aa)
   )
+  (label "Buzz" (at 199.39 121.92 180)
+    (effects (font (size 1.27 1.27)) (justify right bottom))
+    (uuid 3ad0079a-f5c2-4d19-8858-62b1390dcf68)
+  )
   (label "ESTOP_IN" (at 199.39 114.3 180)
     (effects (font (size 1.27 1.27)) (justify right bottom))
     (uuid 4d43f9ce-68c5-4d9d-b5ab-77e8fee90c30)
@@ -2357,6 +2360,10 @@
   (label "TX" (at 246.38 119.38 0)
     (effects (font (size 1.27 1.27)) (justify left bottom))
     (uuid 87d45abe-7694-4cc8-9c0e-e035cdf56258)
+  )
+  (label "Buzz" (at 262.89 139.7 180)
+    (effects (font (size 1.27 1.27)) (justify right bottom))
+    (uuid 88fcb3b2-e22a-45da-ae72-b8bad33cf65b)
   )
   (label "ACT0_FB" (at 35.56 62.23 180)
     (effects (font (size 1.27 1.27)) (justify right bottom))
@@ -3422,7 +3429,7 @@
     (property "Value" "BATT_MON_SELECT" (id 1) (at 200.66 85.0899 0)
       (effects (font (size 1.27 1.27)) (justify left))
     )
-    (property "Footprint" "Connector_PinSocket_2.54mm:PinSocket_1x03_P2.54mm_Vertical" (id 2) (at 198.12 83.82 0)
+    (property "Footprint" "Connector_PinHeader_2.54mm:PinHeader_1x03_P2.54mm_Vertical" (id 2) (at 198.12 83.82 0)
       (effects (font (size 1.27 1.27)) hide)
     )
     (property "Datasheet" "~" (id 3) (at 198.12 83.82 0)
@@ -3955,7 +3962,7 @@
       (reference "J8") (unit 1) (value "I2C") (footprint "Connector_PinHeader_2.54mm:PinHeader_1x02_P2.54mm_Vertical")
     )
     (path "/c3e3b447-1acd-4f6c-a522-0cc6e96c7d76"
-      (reference "J9") (unit 1) (value "BATT_MON_SELECT") (footprint "Connector_PinSocket_2.54mm:PinSocket_1x03_P2.54mm_Vertical")
+      (reference "J9") (unit 1) (value "BATT_MON_SELECT") (footprint "Connector_PinHeader_2.54mm:PinHeader_1x03_P2.54mm_Vertical")
     )
     (path "/c06ee988-6065-4d85-b30f-f7b82e32170d"
       (reference "J10") (unit 1) (value "UART") (footprint "Connector_PinHeader_2.54mm:PinHeader_1x02_P2.54mm_Vertical")
